Fix numpy types

+++ eye_tracking/ptgaze/common/visualizer.py
@@ -22,12 +22,12 @@ class Visualizer:
     def draw_bbox(self, bbox: np.ndarray, color: Tuple[int, int, int] = (0, 255, 0), lw: int = 1) -> None:
         assert self.image is not None
         assert bbox.shape == (2, 2)
-        bbox = np.round(bbox).astype(np.int).tolist()
+        bbox = np.round(bbox).astype(np.int32).tolist()
         cv2.rectangle(self.image, tuple(bbox[0]), tuple(bbox[1]), color, lw)
 
     @staticmethod
     def _convert_pt(point: np.ndarray) -> Tuple[int, int]:
-        return tuple(np.round(point).astype(np.int).tolist())
+        return tuple(np.round(point).astype(np.int32).tolist())
 
     def draw_points(self, points: np.ndarray, color: Tuple[int, int, int] = (0, 0, 255), size: int = 3) -> None:
         assert self.image is not None
@@ -58,7 +58,7 @@ class Visualizer:
         assert face.head_position is not None
         assert face.landmarks is not None
         # Get the axes of the model coordinate system
-        axes3d = np.eye(3, dtype=np.float) @ Rotation.from_euler("XYZ", [0, np.pi, 0]).as_matrix()
+        axes3d = np.eye(3, dtype=np.float32) @ Rotation.from_euler("XYZ", [0, np.pi, 0]).as_matrix()
         axes3d = axes3d * length
         axes2d = self._camera.project_points(axes3d, face.head_pose_rot.as_rotvec(), face.head_position)
         center = face.landmarks[self._center_point_index]
